fix(channels-status-config-summary): split connect-recovery test to satisfy file budget
