+++ test/cli/connect-recovery.test.ts
@@ -1413,97 +1413,4 @@ describe("CLI dispatch", () => {
     expect(log.includes("sandbox get alpha")).toBeTruthy();
     expect(log.includes("sandbox connect alpha")).toBeTruthy();
   });
-
-  it("connect surfaces sandbox-not-found when recovery cannot find the requested sandbox (#2164)", () => {
-    const home = fs.mkdtempSync(
-      path.join(os.tmpdir(), "nemoclaw-cli-connect-unknown-after-recovery-"),
-    );
-    const localBin = path.join(home, "bin");
-    const nemoclawDir = path.join(home, ".nemoclaw");
-    fs.mkdirSync(localBin, { recursive: true });
-    fs.mkdirSync(nemoclawDir, { recursive: true });
-    fs.writeFileSync(
-      path.join(nemoclawDir, "onboard-session.json"),
-      JSON.stringify(
-        {
-          version: 1,
-          sessionId: "session-1",
-          resumable: true,
-          status: "complete",
-          mode: "interactive",
-          startedAt: "2026-03-31T00:00:00.000Z",
-          updatedAt: "2026-03-31T00:00:00.000Z",
-          lastStepStarted: "policies",
-          lastCompletedStep: "policies",
-          failure: null,
-          sandboxName: "alpha",
-          provider: "nvidia-prod",
-          model: "nvidia/nemotron-3-super-120b-a12b",
-          endpointUrl: null,
-          credentialEnv: null,
-          preferredInferenceApi: null,
-          nimContainer: null,
-          policyPresets: null,
-          metadata: { gatewayName: "nemoclaw" },
-          steps: {
-            preflight: { status: "complete", startedAt: null, completedAt: null, error: null },
-            gateway: { status: "complete", startedAt: null, completedAt: null, error: null },
-            sandbox: { status: "complete", startedAt: null, completedAt: null, error: null },
-            provider_selection: {
-              status: "complete",
-              startedAt: null,
-              completedAt: null,
-              error: null,
-            },
-            inference: { status: "complete", startedAt: null, completedAt: null, error: null },
-            openclaw: { status: "complete", startedAt: null, completedAt: null, error: null },
-            policies: { status: "complete", startedAt: null, completedAt: null, error: null },
-          },
-        },
-        null,
-        2,
-      ),
-      { mode: 0o600 },
-    );
-    fs.writeFileSync(
-      path.join(localBin, "openshell"),
-      [
-        "#!/usr/bin/env bash",
-        'if [ "$1" = "status" ]; then',
-        "  echo 'Server Status'",
-        "  echo",
-        "  echo '  Gateway: nemoclaw'",
-        "  echo '  Status: Connected'",
-        "  exit 0",
-        "fi",
-        'if [ "$1" = "gateway" ] && [ "$2" = "info" ]; then',
-        "  echo 'Gateway Info'",
-        "  echo",
-        "  echo '  Gateway: nemoclaw'",
-        "  exit 0",
-        "fi",
-        'if [ "$1" = "sandbox" ] && [ "$2" = "list" ]; then',
-        "  echo 'No sandboxes found.'",
-        "  exit 0",
-        "fi",
-        'if [ "$1" = "--version" ]; then',
-        "  echo 'openshell 0.0.16'",
-        "  exit 0",
-        "fi",
-        "exit 0",
-      ].join("\n"),
-      { mode: 0o755 },
-    );
-
-    const r = runWithEnv("beta connect", {
-      HOME: home,
-      PATH: `${localBin}:${process.env.PATH || ""}`,
-    });
-
-    expect(r.code).toBe(1);
-    expect(r.out.includes("Sandbox 'beta' does not exist")).toBeTruthy();
-    // Recovery from onboard-session.json restores "alpha" into the local registry,
-    // so the helper lists it rather than the empty-registry onboard hint.
-    expect(r.out.includes("Registered sandboxes: alpha")).toBeTruthy();
-  });
 });
